requirements.txt
alembic==1.12.0
anyio==4.0.0
APScheduler==3.10.4
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
humanfriendly==10.0
idna==3.4
iniconfig==2.0.0
Mako==1.2.4
MarkupSafe==2.1.3
packaging==23.1
pluggy==1.3.0
psycopg2==2.9.7
pytest==7.4.2
python-telegram-bot==20.5
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.20
typing_extensions==4.7.1
tzlocal==5.0.1
urllib3==2.0.4
